Add setRequestLogger-method to client-API.
Background: turbomanage http-client will use the default ConsoleRequestLogger (logs everything), if a RequestHandler is supplied to the MailjetClient constructor.
In practice this meant that it was not possible to use a custom handler without logging every request.

This commit fix the issue by exposing a setter for the client's request logger.